Barcode and QR generator/requirements.txt
qrcode[pil]
python-barcode
pillow
pyinstaller
